+++ src/components/posts/AllPosts.tsx
@@ -3,6 +3,7 @@ import { useInfiniteQuery } from "@tanstack/react-query";
 import React, { useEffect, useRef } from "react";
 import postApi from "../../api/postApi";
 import BlueSpinner from "../../common/components/spinners/BlueSpinner";
+import { useVisibleEffect } from "../../hooks/visibleEffect";
 import Post from "./Post";
 
 // 쿼리 로딩 => 온석세스 셋포스트 인밸리드 => 또 쿼리
@@ -26,25 +27,7 @@ const AllPosts = () => {
   });
 
   /* Infinite scroll */
-  const observerTarget = useRef(null);
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting) {
-          fetchNextPage();
-        }
-      },
-      { threshold: 1 }
-    );
-    if (observerTarget.current) {
-      observer.observe(observerTarget.current);
-    }
-    return () => {
-      if (observerTarget.current) {
-        observer.unobserve(observerTarget.current);
-      }
-    };
-  }, [observerTarget]);
+  const { observerTarget } = useVisibleEffect(fetchNextPage);
 
   return (
     <>
