+++ src/components/posts/UsersPosts.tsx
@@ -3,6 +3,7 @@ import { useInfiniteQuery } from "@tanstack/react-query";
 import React, { useEffect, useRef } from "react";
 import postApi from "../../api/postApi";
 import BlueSpinner from "../../common/components/spinners/BlueSpinner";
+import { useVisibleEffect } from "../../hooks/visibleEffect";
 
 import Post from "./Post";
 
@@ -24,25 +25,8 @@ const UsersPosts = ({ uids }: UsersPostsProps) => {
   });
 
   /* Infinite scroll */
-  const observerTarget = useRef(null);
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting) {
-          fetchNextPage();
-        }
-      },
-      { threshold: 1 }
-    );
-    if (observerTarget.current) {
-      observer.observe(observerTarget.current);
-    }
-    return () => {
-      if (observerTarget.current) {
-        observer.unobserve(observerTarget.current);
-      }
-    };
-  }, [observerTarget]);
+  /* Infinite scroll */
+  const { observerTarget } = useVisibleEffect(fetchNextPage);
 
   return (
     <>
